* Moved method to using class
